requirements.txt
PySide6>=6.7.0
qrcode[pil]>=7.4.2
python-barcode>=0.15.1
Pillow>=10.0.0
openpyxl>=3.1.2
pywin32>=306; platform_system == "Windows"
